+++ excel_to_json/urls.py
@@ -18,6 +18,8 @@ from django.contrib import admin
 from django.urls import path, include
 
 urlpatterns = [
-    path('admin/', admin.site.urls),path('api/', include ('api.urls')),
+    path('admin/', admin.site.urls),
+    path('api/', include('api.urls')),
     path('api/pdf/', include('pdfconvert.urls')),
+    path('api/ocr/', include('ocr.urls')),
 ]
