+++ ocr/__init__.py
@@ -1,0 +1,3 @@
+from .parsers import TextractOCRParser
+__all__ = ['TextractOCRParser']
+
